Added enemy generation. - changes based on level, been changing variable name a lot and moving things to test so I'll need to check over enemy spawn behaviour - I let the IDE set up bullet collision for me quickly to work on the respawning mechanic but, I'd like to take a closer look at my hit boxes
Included a direction component when assesing if an item needs to be transfered to the opposite side, allowing for items to exist on the edge, without getting infinitely trapped

Added a basic start game screen with a button to start the actual game

Added a setInterval on my requestAnimationFrame to try and keep the frames steady. Want to be using timestamps but I'll need to find time for the research.